Simplify function types for HandlerFutures
We have a trait implementation for functions that take a response and
return a `Box<HandlerFuture>` which we've been relying on to support
basic functions as `Handler` implementations.

However we also have an implementation for `IntoHandlerFuture` for
functions for `hyper::server::Response`.

By combining these, as proposed here, we can take simple function types
from `Request -> Box<HandlerFuture>` to simply `Request -> Response`
which is going to be a lot easier for newcomers.