Refactor to support async middleware
After a fair bit of wrestling with the borrow checker, the most
"correct" implementation I could come up with was to introduce
`NewHandler` and `NewMiddleware` traits, analogous to tokio's
`NewService`.

In order for middleware to be asynchronous, the middleware pipeline
needs to be spawned per-request, because the "chain" representing the
tail of the pipeline must be moved into a future in order for the chain
to be called there. So, calling the pipeline must consume it.

This will have flow-on effects to the rest of the project, since we
should keep to a single "pattern" for spawning handlers.