Initial commit of Node
This is a recursive member of what will be our route Tree. It is
essentially a container but includes smarts within it's traverse method
that allows it to determine if it matches an individual segment of a
request path and when it does offer potential routes to the router for
further consideration.