Initial cut of Tree backed Router
This still needs further enhancement but is shipping to PR earlier than
usual given the large amount of change that has already been introduced
and the changes that still need to be made once this is locked down with
the rest of the team.

Notably:

- Pipelines are not currently supported, refinement needs to be
considered to allow them to be stored and executed in a generic manner
regardless of middleware used

- Path segment variables are not currently being turned into attributes
in the `State` object

- The absolute failure state of `future:err` needs to be logged, likely
at a higher level in the application than Router specifically but this
needs to be decided.