Refine Tree segment handling
So the outside world sees a "correct" segment structure that includes a
segment identifying the root node, "/", the code here had to be slightly
adjusted.

The Tree also now exposes a function for splitting the request path so
that if we need to tweak this again in future it won't be externally
breaking (as opposed to others implementing `split` processing directly).