Simplify Tree structure and support globing
Previously we had an implementation that added complexity that ended up
not really being necessary and limited our ability to support more
advanced concepts such as Route Globbing.

This commit addresses that by:

* Properly considers "/" as the root segment in a request path;
* Limiting Nodes to have 1 of 4 specific types of segment;
* Sorting child nodes within a Tree to ensure that more specific segments matches are
always attempted first;
* Reducing the number of passes we need to do in the Tree from 2 to 1 by
extracting segment match data as we undertake the path build process.

Further refinements will still be made here, in particular spliting
current code into a `TreeBuilder` + `Tree` and `NodeBuilder`+`Node` to
ensure that finalization is complete before a `Router` is able to access
a `Tree` and to remove any ability to further modify structures.