Implement Request path extraction in type safe way
Provides a custom derive `RequestPathExtractor` for client applications
to keep implementation effort to a minimum.

Also defines a new trait `FromRequestPath` that translates String to
required values. Default implementations make use of the existing
`FromStr` trait to save re-implementation.

Optional values may be used so the same Struct can be used for multiple
Routes but this should be a last resort.

Examples of use, including Optional segmnts, have been added to
KitchenSink.

RouteImpl will likely be changed as the last part of this series of
commits to split out the RequestPathExtractor (and future extractors,
such as those for query string) into a seperate struct. This should
remove the need for the current
`RouteImpl<_, _, _, _, NoopRequestPathExtractor>`
type code which is not quite ideal from a readabilty perspective.