Revert Tree to take &str for traverse
One of the biggest downsides for using URI with Tree::traverse was the
need to convert data into String at multiple points, which is not ideal
for a core part of response generation.

By reverting to traverse taking &str, which is already percent decoded by
the Router, we can avoid a lot of these unnecessary
allocations/conversions.