Ensure the name of query string pairs is decoded
This is required by spec. There is some potential for this to be a
performance issue given that FormUrlDecoded is now acting as a Key to the
backing HashMap for QueryStringMapping.

Might be worth thinking about this HashMap (and potentially that of
SegmentMapping) returning to simply use str for keys and values, the
contract for encoding would then become implied by the Mapping types but
as only Gotham code can instantiate these that may be reasonable.