Alter RequestMatcher is_match to return HTTP error
The Router needs to be able to respond to clients with more specific
error codes than then 404 or 500 possible with the temporary boolean API.

With this change the Router will be able to signal a more specific error
case when no Routes are able to be matched, e.g. 405 method not allowed.