Refactor Router handler and add ResponseExtenders
Starts to finalize the implementation we'll have for Router, at least
at 0.1.

Adds an additional ability to Router via the RequestExtender and Extender
types that allows code to make a final change to a response generated
elsewhere within the application prior to it being provided to a client.

The initial use case here is error message formatting for the 400
and 500 range but is implemented generically so as to support
enhancements for any status code.