Reference pipelines from Dispatcher not Router
Previously we assigned the complete set of all created Pipeline instances
to Router when it was constructed.

This led to complexity in the presented API due to the requirement to
then know, in advance of router construction the type of all Pipelines
stored within the associated BorrowBag instance which is not possible.

Practially speaking you could not achieve something like this:

```
pub fn make_me_a_router() -> Router {}
```

Whilst reviewing the implementations it was clear that only the
Dispatcher really ever needs access to the complete set of Pipelines
which it uses through call methods on PipelineHandlerChain.

The solution here has to be to re-develop Dispatcher as a Trait Object
allowing Boxed instances of DispatcherImpl to be assigned to a RouteImpl
during construction.

This change has had the wonderful side effect of allowing us to remove the
Pipeline type parameter from Router, RouterData, Tree, Node, Path,
and RouteImpl.