Add type alias PipelineSet and supporting items
The distinction between Pipelines, PipelineChainHandlers and BorrowBag
was confusing previously and made the API onerous on the developer.

This introduces the type aliases:

* EditablePipelineSet - A set of `Pipeline` instances that is still being
extended, generally for use within `Router` setup code.
* PipelineSet - A complete set of possible `Pipeline` instances for use
with a `Router` wrapped in an `Arc`.

They've been documented and integrated to tests and kitchen-sink to show
intended usage.