Initial support for delegated Routers
This currently relies on a few things that are sub-optimal especially
that of identifying a segment as being considered 'delegated' when the
correct view is the Node itself represents a delegation for the segment
it holds. This will be corrected in a subsequent commit.

Extensively extends RequestPathSegments to allow for the delegated case.

Delegated Router instances will continue to believe that they've accepted
the Request at the root of the Request path. They will not be provided
access to any segment mapping data that a parent Router(s) may have been
able to extract. For this reason, it is recommended that paths which lead
to a delegation Node use static segments only, however other
segment types will happily function.

Recursive delegation of Routers should be possible but a depth of greater
than 1 has not (yet) been proven.

Test cases are still to be provided here, awaiting the changes
anticipated for Node.