Invoke NoopExtender when Response body not present
This was previously waiting on a change to Hyper which I've recently
pushed a PR for.

This code is redundant in the NoopExtender but serves as a reasonable
basis for how proper extenders should be implemented in the future, that
is only be invoked if nothing else has previously set the Response body.