Process QueryString extraction on every Request
Prior to this the Router would bypass query string extraction and mapping
if there was no query string associated with the Request.

This was not correct for Routes that have dictated that a specific set of
name+val params must be available in order for the Request to be validly
handled (i.e. Non Optional values in a struct implementing
QueryStringExtractor).

In addition an invalid set of name+val params results in the Router
responding with 400 Bad Request now, instead of 500 Internal Server
Error.

Extended trace logging to make working with query strings from
applications easier to debug.