Refact ResponseExtender to be more applicable
Previously the ResponseExtender focused only on modifications to the
final response which was not flexible enough.

This change to the structure of Gotham allows response extension
potentially handled by multiple components associated with a Route.

THe previous functionality of ResponseExtender is now exposed through the
ResponseFinalizer which now invokes a ResponseExtender, if present, for
the status code associated with the Response.