Allow RequestPathExtractor to extend Response
In the case of an error RequestPathExtractors might need to place custom
data into the Response instead of simply relying on Router to provide a
400 Bad Request. e.g. Adding specific JSON to the Response body or
setting an expected ContentType.

To ease implementation in the generic case the default derive,
`PathExtractor` continues to simply set 400 Bad Request on the response
in the case of error. The new derive `BasePathExtractor` is provided
for cases where custom Response data is going to be implemented
by applications.