Introduce the FromState trait
For use by structs, such as those created by extractors, who know that
their existence in State is mandated by the contracts provided by the
Gotham type system and Router implementation.

May also be implemented when it is known that other sources, such as a
Middleware, will have previously populated State with data before
allowing request processing to proceed.

In the event that State does not hold the type a panic is created and the
thread will be torn down (this may be later enhanced with panic catching
middleware to ensure the system stays alive in such a situation).