Fix behaviour for session deserialize failures
As noted in the TODO, this would fail for applications that changed
their session struct but had persistent session data that survived a
restart. Falling back to a new session seems like the only sensible way
to handle it.

The `WARN` log entry ought to help with diagnostics if this is triggered
another way.