Compare SegmentType when comparing segments
When building a `Router` one route at a time (instead of building the
tree hierarchy in order), there's a need to mutably traverse the
`TreeBuilder` / `NodeBuilder` structure. When lazily adding a segment,
we want to ensure we're creating a segment with the correct
`SegmentType`, so we need to compare `SegmentType` on traversal to
ensure we don't have a false match.