Support Windows via socket queue implementation
Attempting to use cloned sockets on Windows resulted in a lot of errors
from Tokio's TcpListener, and no successful responses. Exploration of
how to concurrently serve a single TcpListener on Windows was
unsuccessful, leading to the solution here where a single thread accepts
connections and then dispatches them via a queue to the per-thread Tokio
reactors.