Implement regex matching for request path segments
Adds anchored (^..$) regex matching to the request path segment handling
and drops the `unimplemented!()` macro in favor of a concrete
implementation.

This also adds an `OrderedRegex` newtype as the exist Regex didn't
implement the traits required for PartialEq, Eq, PartialOrd, and Ord.
See the following Github issues for reasons why Regex doesn't natively
implement this:
https://github.com/rust-lang/regex/issues/313
https://github.com/rust-lang/regex/issues/364

This (hopefully) addresses and closes #10.