Fix blank session cookie bug
This looks like a refactor gone wrong in b0b39047, which was missed
and got merged after the initial (pre-refactor) version had already
been reviewed. The problem is that an empty `Vec` with a capacity is
still seen as a 0-length slice, since none of the elements are
initialized.

The RNG attempts (quite successfully) to fill the 0-length slice, and
the session is established with a 0-length identifier.