Refine SessionCookieConfig implementation
The previous implementation was not ideal when we came to implement
configurable path values, mostly in that it was complex to update an
option that had runtime data associated with it.

The new implementation simply sets values on the SessionCookieConfig
struct directly making updates easier and future extensions simpler.

External usage should not be impacted by this change as the
existing public API was able to be kept and in the case of
`anchor_to_path` extended.