Add top-level catch_unwind to stop process death
When something panics in a Gotham application, we can trap the panic at
the `NewHandlerService` level and respond with a blank 500 error
response. This is intended to be a last-ditch attempt at stopping the
application falling over, and panics should instead be handled in a
couple of more intelligent ways:

1. Applications should not cause a panic for anything but the most dire
   of runtime errors; and
2. A panic-catching middleware will be introduced which attempts to
   handle these panics more gracefully.

Notably, this logic **does not** prevent panics occurring if the
returned `Future` panics.

Ref #17