Support RefUnwindSafe requirement on middleware
The latest Gotham release adds a requirement for middleware to be
`RefUnwindSafe` in order to allow a panic within Middleware and Handlers
to be represented as an internal server error.

Unfortunately the r2d2::Pool and other types used within the pool
hierachy are not unwind safe. We've worked around this by:

    1. Reducing the surface area to only the clone function, via
    splitting of our types into DieselMiddleware and DieselMiddleware and
    2. Trapping a panic within clone, and aborting the process to avoid
    leaving invalid state in the mutable storage within r2d2::Pool et al.