Implement Clone for DieselMiddleware
We need the ability to pass a single DieselMiddleware instance to
multiple router instances for Modular Gotham applications which is
enabled by this commit.

Supporting this ensures that the Diesel Middleware can me made part of
pipelines where it is necessary thus not being traversed
application wide e.g. endpoints/controllers that have no need for
database interaction.