Write ThreadId as debug log statement
To assist with diagnosing issues with concurrency, we can ensure the
current thread is written as a debug log statement before the request is
dispatched to a Handler. The value should be considered opaque and need
not correspond to any system-designated identifier for the thread, but
is sufficient for proving whether requests are served by the same, or
different threads.