Remove lifetime from TestResponse
In 3rd party code it may be useful to setup a TestServer and TestResponse
within a common method, returning both to the the test harness for
further evaluation.

With lifetimes as specified this was a lot harder to achieve and meant
tests violated DRY to make this happen.

These changes specifically address this use case via an Rc, thus
eliminating the need for a lifetime specifier.