Update Gotham to 0.2-dev
In the future appening `-dev` to master will occur the first time we
start to make additions towards a future release (for any crates).

Versions with `-dev` will not get published to crates.io.

Once a set of features for a release are complete, `-alpha` and `-beta`
identifiers and publishing to crates.io will occur, eventually followed
by the final release.