Refine RouterBuilder support for custom Matchers
The previous implementation required the `request` and `explicit_request`
methods in order to support the existing usage of `vec![METHOD]` in the
builder as well the new ability to provide custom matchers.

During PR it was noted that `IntoRouteMatcher` would allow unification
around a singular `request` method once again, which is how we've
arrived at the much more useful implementation in this commit.