Allow builder to extend current route matcher
As part of #177 we allowed the request method to take anything which is
an `IntoRouteMatcher` which is useful for cases where we want to be very
specific about the matchers in use for a given route.

However this leads to code like this:

```
  let methods = vec![Method::Get, Method::Head];
  let supported_media_types = vec![mime::APPLICATION_JSON];
  let method_route_matcher = MethodOnlyRouteMatcher::new(methods);
  let accept_header_route_matcher = AcceptHeaderRouteMatcher::new(supported_media_types);
  let matcher = AndRouteMatcher::new(method_route_matcher, accept_header_route_matcher);
  route
      .request(matcher, "/")
      .to(my_handler);
```

It would be nicer for this common situation to not have to manually
specify the `MethodOnlyRouteMatcher` and the `AndRouteMatcher`.

This commit supports this case which means we end up with a much nicer
router definition:

```
  let supported_media_types = vec![mime::APPLICATION_JSON];
  let accept_header_route_matcher = AcceptHeaderRouteMatcher::new(supported_media_types);
  route
      .get_or_head("/")
      .add_route_matcher(accept_header_route_matcher)
      .to(my_handler);
```