bind_server: do not exit on protocol error
As part of #370, this function was updated from spawning a future to
awaiting it through `try_for_each_concurrent`.

As as a result of that change, any protocol error will result in
terminating the server. This can happen if a client is connecting using
HTTP instead of HTTPS, or even the client rejecting the server's HTTPS
certificate and informing the server of said rejection!

To reproduce this, you can start the hello_world_tls example, and send a
request using plain HTTP. The server will immediately exit. With this
patch applied, it does not.