assets: add range request support for FileHandler
This adds range request support for FileHandler. This is a critical
feature for handling large files, or any application that will require
range request support, e.g. video or audio streaming.

This implementation currentyl only supports ranges that result in a
single response, i.e. multiple ranges are not supported.

This implementation will return an error if a range is open from both
ends, or if the range beginning is larger than the end.

Signed-off-by: Janne Pelkonen <jpelkonen@ideanovatech.com>